only delete old models for models that have already run

+++ model_processing/get_models.py
@@ -523,9 +523,10 @@ conn.commit ()
 set_model_to_waiting (model_name)
 log ("Model processing completed successfully.".format(fmt_timestep),"INFO", model_name)
 
-log ("Deleting old model run (rasters.{0}_{1}).".format (model_name,str(last_checked_total_seconds - config["maxModelAge"])), "INFO", model_name)
-cur.execute ("DROP TABLE rasters.{0}_{1}".format (model_name,str(last_checked_total_seconds - config["maxModelAge"])))
-conn.commit ()
+if last_checked_total_seconds > 0:
+    log ("Deleting old model run (rasters.{0}_{1}).".format (model_name,str(last_checked_total_seconds - config["maxModelAge"])), "INFO", model_name)
+    cur.execute ("DROP TABLE rasters.{0}_{1}".format (model_name,str(last_checked_total_seconds - config["maxModelAge"])))
+    conn.commit ()
 
 log ("Cleaning up the logs...", "INFO")
 cur.execute ("DELETE FROM logging.processing_logs WHERE timestamp < now() - interval '7 days'")
